ajuste de comentarios
