+++ src/main/java/com/compiler/lexical/token/TokenType.java
@@ -16,7 +16,6 @@ public enum TokenType {
     LEFT_PAREN,
     RIGHT_PAREN,
     NUMBER,
-    COMMA,
     INT,
     FLOAT,
     PRINT,
